Removing state handling from audio_classifier and giving gpio_handler a job
Removed the HTTP request as well as the GPIO responsibilities from audio_classifier

Created gpio_handler to deal with the logic of state. In order for the pin connecting to the jetson to output high, there must be a voltage given to GPIO 26. Without this voltage, the output voltage is forced to zero. This acts as an override switch.
